% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{check_status}
\alias{check_status}
\title{Check Status of Upload/Validation Ticket}
\usage{
check_status(ticket)
}
\arguments{
\item{ticket}{A character.}
}
\value{
A character.
}
\description{
Check Status of Upload/Validation Ticket
}
\examples{
\dontrun{
check_status("testuser_20180523_19544")
}
}
\references{
\url{http://docs.immport.org/#API/DataUploadAPI/datauploadapi/#status-of-upload-ticket-with-authentication}
}
