% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{download_ticket_report}
\alias{download_ticket_report}
\title{Download Database Report on a Upload/Validation Ticket}
\usage{
download_ticket_report(ticket, output_dir = ".")
}
\arguments{
\item{ticket}{A character.}

\item{output_dir}{A character. Output directory.}
}
\value{
A character. File path to the report.
}
\description{
Download Database Report on a Upload/Validation Ticket
}
\examples{
\dontrun{
# download the database report of a upload/validation ticket
download_ticket_report("testuser_20180523_19544")

# inspect the database report
file.edit("testuser_20180523_19544_uploadReport.txt")
}

}
\references{
\url{http://docs.immport.org/#API/DataUploadAPI/datauploadapi/#database-information-on-upload-ticket-request-with-authentication}
}
