% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{get_ticket_report}
\alias{get_ticket_report}
\title{Retrieve Database Report on a Upload/Validation Ticket}
\usage{
get_ticket_report(ticket)
}
\arguments{
\item{ticket}{A character.}
}
\value{
A character. Raw text of the report.
}
\description{
Retrieve Database Report on a Upload/Validation Ticket
}
\examples{
\dontrun{
get_ticket_report("testuser_20180523_19544")
}

}
\references{
\url{http://docs.immport.org/#API/DataUploadAPI/datauploadapi/#database-information-on-upload-ticket-request-with-authentication}
}
