% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{get_ticket_summary}
\alias{get_ticket_summary}
\title{Retrieve Summary Information on Upload/Validation Ticket}
\usage{
get_ticket_summary(ticket)
}
\arguments{
\item{ticket}{A character.}
}
\value{
A list.
}
\description{
Retrieve Summary Information on Upload/Validation Ticket
}
\examples{
\dontrun{
get_ticket_summary("testuser_20180523_19544")
}

}
\references{
\url{http://docs.immport.org/#API/DataUploadAPI/datauploadapi/#summary-information-on-upload-ticket-request-with-authentication}
}
