% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImmPortR-package.R
\docType{package}
\name{ImmPortR-package}
\alias{ImmPortR}
\alias{ImmPortR-package}
\title{ImmPortR: An R wrapper Around the ImmPort API}
\description{
Wraps around the ImmPort API to query datasets from ImmPort Shared Data and upload data to ImmPort Private Data, and utilizes the Aspera CLI to download files from ImmPort Shared Data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RGLab/ImmPortR}
  \item Report bugs at \url{https://github.com/RGLab/ImmPortR/issues}
}

}
\author{
\strong{Maintainer}: Ju Yeong Kim \email{jkim2345@fredhutch.org}

}
\keyword{internal}
