% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspera.R
\name{list_immport}
\alias{list_immport}
\title{List file or directory from ImmPort}
\usage{
list_immport(path)
}
\arguments{
\item{path}{A character. File or directory to list}
}
\value{
A list.
}
\description{
List file or directory from ImmPort
}
\examples{
\dontrun{
list_immport("/SDY1/StudyFiles")
}
}
\references{
\url{http://docs.immport.org/#API/FileDownloadAPI/filedownloadapi/#example-for-listing-a-shared-data-directory-or-file}
}
