% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{list_workspaces}
\alias{list_workspaces}
\title{Retreive Set of Workspaces}
\usage{
list_workspaces()
}
\value{
A list. Set of workspaces on which a user can perform and upload or
validation.
}
\description{
Retreive Set of Workspaces
}
\examples{
\dontrun{
list_workspaces()
}

}
\references{
\url{http://docs.immport.org/#API/DataUploadAPI/datauploadapi/#set-of-workspaces-request-with-authentication}
}
