% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query_datarelversion}
\alias{query_datarelversion}
\title{query the latest study data release version}
\usage{
query_datarelversion()
}
\value{
A character.
}
\description{
query the latest study data release version
}
\examples{
\dontrun{
query_datarelversion()
}
}
\references{
\url{http://docs.immport.org/#API/DataQueryAPI/dataqueryapi/}
}
