% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query_dataset}
\alias{query_dataset}
\title{query dataset by study accession}
\usage{
query_dataset(study, dataset)
}
\arguments{
\item{study}{a character. study accession.}

\item{dataset}{a character. dataset name. Datasets are "elisa", "elispot",
"fcsAnalyzed", "hai", "hlaTyping", "kirTyping", "mbaa", "neutAbTiter", "pcr"}
}
\value{
data.frame
}
\description{
query dataset by study accession
}
\examples{
\dontrun{
elisa <- query_dataset("SDY269", "elisa")
elispot <- query_dataset("SDY269", "elispot")
fcsAnalyzed <- query_dataset("SDY269", "fcsAnalyzed")
hai <- query_dataset("SDY269", "hai")
hlaTyping <- query_dataset("SDY269", "hlaTyping")
kirTyping <- query_dataset("SDY269", "kirTyping")
mbaa <- query_dataset("SDY269", "mbaa")
neutAbTiter <- query_dataset("SDY269", "neutAbTiter")
pcr <- query_dataset("SDY269", "pcr")
}
}
