% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query_filePath}
\alias{query_filePath}
\title{query file path by study acession}
\usage{
query_filePath(study)
}
\arguments{
\item{study}{a character. study accession.}
}
\value{
A data.frame.
}
\description{
query file path by study acession
}
\examples{
\dontrun{
query_filePath("SDY269")
}
}
\references{
\url{http://docs.immport.org/#API/DataQueryAPI/dataqueryapi/#example-of-a-request-on-the-filepath-endpoint}
}
