% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query_findAllStudyAccessions}
\alias{query_findAllStudyAccessions}
\title{query a list of all study accessions}
\usage{
query_findAllStudyAccessions()
}
\value{
A vector.
}
\description{
query a list of all study accessions
}
\examples{
\dontrun{
query_findAllStudyAccessions()
}
}
\references{
\url{http://docs.immport.org/#API/DataQueryAPI/dataqueryapi/}
}
