% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query_study_summary}
\alias{query_study_summary}
\title{query the study summary}
\usage{
query_study_summary(study)
}
\value{
A list
}
\description{
query the study summary
}
\examples{
\dontrun{
query_study_summary()
}
}
\references{
\url{http://docs.immport.org/#API/DataQueryAPI/dataqueryapi/}
}
