% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{validate_zip}
\alias{validate_zip}
\title{Validate a Zip File}
\usage{
validate_zip(file_path, workspace_id, upload_notes = "")
}
\arguments{
\item{file_path}{A character. File path to the zip file to upload.}

\item{workspace_id}{An integer.}

\item{upload_notes}{A character. Optional.}
}
\value{
A character. Ticket.
}
\description{
Validate a Zip File
}
\examples{
\dontrun{
validate_zip("groundbreakingStudy.zip")
}
}
\references{
\url{http://docs.immport.org/#API/DataUploadAPI/datauploadapi/#zip-file-upload-for-validation-with-authentication}
}
