% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspera.R
\name{download_immport}
\alias{download_immport}
\title{Downlaod file or directory from ImmPort}
\usage{
download_immport(path, output_dir = ".")
}
\arguments{
\item{path}{A character. File or directory to download.}

\item{output_dir}{A character. Output directory.}
}
\description{
Downlaod file or directory from ImmPort
}
\examples{
\dontrun{
download_immport("/SDY1/StudyFiles/Casale_Study_Summary_Report.doc")
}
}
