% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query_dataset}
\alias{query_dataset}
\title{query dataset by study accession}
\usage{
query_dataset(dataset, study)
}
\arguments{
\item{dataset}{a character. dataset name.}

\item{study}{a character. study accession.}
}
\value{
list
}
\description{
query dataset by study accession
}
\examples{
\dontrun{
elisa <- query_dataset("elisa", "SDY269")
elispot <- query_dataset("elispot", "SDY269")
fcsAnalyzed <- query_dataset("fcsAnalyzed", "SDY269")
hai <- query_dataset("hai", "SDY269")
hlaTyping <- query_dataset("hlaTyping", "SDY269")
kirTyping <- query_dataset("kirTyping", "SDY269")
mbaa <- query_dataset("mbaa", "SDY269")
neutAbTiter <- query_dataset("neutAbTiter", "SDY269")
pcr <- query_dataset("pcr", "SDY269")
}
}
